Multiple code improvements - squid:S1149, squid:S1118, squid:S1213, squid:S2259, squid:S1905, squid:S00122

+++ src/main/java/com/darkprograms/speech/microphone/MicrophoneAnalyzer.java
@@ -173,7 +173,7 @@ public class MicrophoneAnalyzer extends Microphone {
 	private double[] applyHanningWindow(double[] signal_in, int pos, int size){
 		for (int i = pos; i < pos + size; i++){
 			int j = i - pos; // j = index into Hann window function
-			signal_in[i] = (double)(signal_in[i] * 0.5 * (1.0 - Math.cos(2.0 * Math.PI * j / size)));
+			signal_in[i] = (signal_in[i] * 0.5 * (1.0 - Math.cos(2.0 * Math.PI * j / size)));
 		}
 		return signal_in;
 	}
